Be more robust against missing dependency analyses.
The conversion worked fine, as long as there is a dependency parse
(empty or non-empty) for every sentence. However, the sentences and
parses could go out of sync if not.

Of course, such errors would quickly show up, because the conversion would
try to access the dependency layer out of bounds.